move server_status into utilities
