+++ instrumentation/opentelemetry-instrumentation-grpc/src/opentelemetry/instrumentation/grpc/_aio_server.py
@@ -17,9 +17,9 @@ import grpc.aio
 import wrapt
 
 from opentelemetry.semconv.trace import SpanAttributes
-from opentelemetry.trace.status import Status, StatusCode
 
 from ._server import OpenTelemetryServerInterceptor, _wrap_rpc_behavior
+from ._utilities import _server_status
 
 
 # pylint:disable=abstract-method
@@ -36,7 +36,7 @@ class _OpenTelemetryAioServicerContext(wrapt.ObjectProxy):
         self._self_active_span.set_attribute(
             SpanAttributes.RPC_GRPC_STATUS_CODE, code.value[0]
         )
-        status = self._server_status(code, details)
+        status = _server_status(code, details)
         self._self_active_span.set_status(status)
         return await self.__wrapped__.abort(code, details, trailing_metadata)
 
@@ -47,33 +47,16 @@ class _OpenTelemetryAioServicerContext(wrapt.ObjectProxy):
             SpanAttributes.RPC_GRPC_STATUS_CODE, code.value[0]
         )
         if code != grpc.StatusCode.OK:
-            status = self._server_status(code, details)
+            status = _server_status(code, details)
             self._self_active_span.set_status(status)
         return self.__wrapped__.set_code(code)
 
     def set_details(self, details):
         self._self_details = details
         if self._self_code != grpc.StatusCode.OK:
-            status = self._server_status(self._self_code, details)
+            status = _server_status(self._self_code, details)
             self._self_active_span.set_status(status)
         return self.__wrapped__.set_details(details)
-
-    def _server_status(self, code, details):
-        error_status = Status(
-            status_code=StatusCode.ERROR, description=f"{code}:{details}"
-        )
-        status_codes = {
-            grpc.StatusCode.UNKNOWN: error_status,
-            grpc.StatusCode.DEADLINE_EXCEEDED: error_status,
-            grpc.StatusCode.UNIMPLEMENTED: error_status,
-            grpc.StatusCode.INTERNAL: error_status,
-            grpc.StatusCode.UNAVAILABLE: error_status,
-            grpc.StatusCode.DATA_LOSS: error_status,
-        }
-
-        return status_codes.get(
-            code, Status(status_code=StatusCode.UNSET, description="")
-        )
 
 
 class OpenTelemetryAioServerInterceptor(
