+++ instrumentation/opentelemetry-instrumentation-grpc/src/opentelemetry/instrumentation/grpc/_utilities.py
@@ -14,6 +14,8 @@
 
 """Internal utilities."""
 
+from opentelemetry.trace.status import Status, StatusCode
+
 
 class RpcInfo:
     def __init__(
@@ -31,3 +33,21 @@ class RpcInfo:
         self.request = request
         self.response = response
         self.error = error
+
+
+def _server_status(code, details):
+    error_status = Status(
+        status_code=StatusCode.ERROR, description=f"{code}:{details}"
+    )
+    status_codes = {
+        grpc.StatusCode.UNKNOWN: error_status,
+        grpc.StatusCode.DEADLINE_EXCEEDED: error_status,
+        grpc.StatusCode.UNIMPLEMENTED: error_status,
+        grpc.StatusCode.INTERNAL: error_status,
+        grpc.StatusCode.UNAVAILABLE: error_status,
+        grpc.StatusCode.DATA_LOSS: error_status,
+    }
+
+    return status_codes.get(
+        code, Status(status_code=StatusCode.UNSET, description="")
+    )
